2.0.5

+++ src/main/java/net/liukrast/eg/content/logistics/board/StringPanelBehaviour.java
@@ -137,7 +137,7 @@ public class StringPanelBehaviour extends AbstractPanelBehaviour implements Cach
             return;
         value = result;
         try {
-            intValue = Integer.parseInt(result);
+            intValue = (int) Float.parseFloat(result);
         } catch (NumberFormatException ignored) {
             intValue = 0;
         }
